refactor : merge 내용에 맞도록 feature 폴더구조 변경 (#96)
component들 ui 폴더로 이동 및 index.ts 수정
