--- src/features/clinic/index.ts
@@ -1,10 +1,0 @@
-export * from './ClinicClientPage'
-export * from './info/ClinicInfoClientPage'
-export * from './info/ClinicInfoModal'
-export * from './ScheduleItem'
-export * from './ScheduleSection'
-export * from './write/AlarmBottomSheet'
-export * from './write/ClinicForm'
-export * from './write/ClinicWriteClientPage'
-export * from './write/DateBottomSheet'
-export * from './write/TagBottomSheet'
