+++ src/features/clinic/ui/ClinicInfoModal.tsx
