+++ src/features/clinic/ui/ScheduleItem.tsx
