+++ src/features/clinic/ui/ScheduleSection.tsx
