+++ src/features/index.ts
@@ -1,3 +1,4 @@
+// auth
 export * from './DashBoardTemplate'
 export * from './sign-in/SignInStep'
 export * from './sign-up/AccountStep'
@@ -5,3 +6,15 @@ export * from './sign-up/CompleteStep'
 export * from './sign-up/SignUpFormProvider'
 export * from './sign-up/TosStep'
 export * from './sign-up/UserInfoStep'
+
+// clinic
+export * from './clinic/ui/AlarmBottomSheet'
+export * from './clinic/ui/ClinicClientPage'
+export * from './clinic/ui/ClinicForm'
+export * from './clinic/ui/ClinicInfoClientPage'
+export * from './clinic/ui/ClinicInfoModal'
+export * from './clinic/ui/ClinicWriteClientPage'
+export * from './clinic/ui/DateBottomSheet'
+export * from './clinic/ui/ScheduleItem'
+export * from './clinic/ui/ScheduleSection'
+export * from './clinic/ui/TagBottomSheet'
